+++ stories/dialogs.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { Store, State } from '@sambego/storybook-state';
-import { select } from '@storybook/addon-knobs/react';
-import { IconWarningBadgedMediumOutline, IconCheckmarkMediumOutline } from '@teamleader/ui-icons';
-import { Banner, Box, Button, ButtonGroup, Dialog, Heading3, TextBody } from '../src';
+import { boolean, number, select, text } from '@storybook/addon-knobs/react';
+import { IconWarningBadgedMediumOutline } from '@teamleader/ui-icons';
+import { Banner, Box, Button, ButtonGroup, COLORS, Dialog, DialogBase, Heading3, TextBody } from '../src';
 
 const sizes = ['small', 'medium', 'large', 'fullscreen'];
 
@@ -13,6 +13,7 @@ const store = new Store({
 
 const handleActiveToggle = () => {
   store.set({ active: !store.get('active') });
+  console.log('handleActiveToggle');
 };
 
 storiesOf('Dialogs', module)
@@ -21,81 +22,58 @@ storiesOf('Dialogs', module)
       propTablesExclude: [Box, Button, ButtonGroup, Banner, TextBody, Heading3, State],
     },
   })
-  .add('Basic', () => (
-    <Box>
-      <Button onClick={handleActiveToggle} label="Open a dialog" />
-      <State store={store}>
-        <Dialog
-          active={false}
-          onEscKeyDown={handleActiveToggle}
-          onOverlayClick={handleActiveToggle}
-          size={select('Size', sizes, 'medium')}
-        >
-          <Banner color="neutral" fullWidth onClose={handleActiveToggle}>
-            <Heading3>Dialog title</Heading3>
-          </Banner>
-          <Box padding={4}>
+  .add('DialogBase', () => {
+    return (
+      <Box>
+        <Button onClick={handleActiveToggle} label="Open a dialog base" />
+        <State store={store}>
+          <DialogBase
+            backdrop={select('backdrop', ['dark'], 'dark')}
+            onEscKeyDown={handleActiveToggle}
+            onOverlayClick={handleActiveToggle}
+            size={select('size', sizes, 'medium')}
+          >
+            <Box padding={4}>
+              <TextBody>Here you can add arbitrary content.</TextBody>
+            </Box>
+          </DialogBase>
+        </State>
+      </Box>
+    );
+  })
+  .add('Dialog', () => {
+    const passHeaderIcon = boolean('pass a headerIcon', false);
+
+    return (
+      <Box>
+        <Button onClick={handleActiveToggle} label="Open a dialog" />
+        <State store={store}>
+          <Dialog
+            headerColor={select('headerColor', COLORS, 'neutral')}
+            headerIcon={passHeaderIcon ? <IconWarningBadgedMediumOutline /> : null}
+            headingLevel={number('headingLevel', 3, { min: 2, max: 3 })}
+            onCloseClick={handleActiveToggle}
+            primaryAction={{
+              label: text('primaryAction.label', 'Confirm'),
+              onClick: () => console.log('primaryAction.onClick'),
+            }}
+            secondaryAction={{
+              label: text('secondaryAction.label', 'Cancel'),
+              onClick: () => console.log('secondaryAction.onClick'),
+            }}
+            tertiaryAction={{
+              children: text('tertiaryAction.children', 'Read more'),
+              onClick: () => console.log('tertiaryAction.onClick'),
+            }}
+            title={text('title', 'Dialog title')}
+            backdrop={select('backdrop', ['dark'], 'dark')}
+            onEscKeyDown={handleActiveToggle}
+            onOverlayClick={handleActiveToggle}
+            size={select('size', sizes, 'medium')}
+          >
             <TextBody>Here you can add arbitrary content.</TextBody>
-          </Box>
-          <ButtonGroup justifyContent="flex-end" padding={4}>
-            <Button label="Cancel" />
-            <Button label="Confirm" level="primary" />
-          </ButtonGroup>
-        </Dialog>
-      </State>
-    </Box>
-  ))
-  .add('succes style', () => (
-    <Box>
-      <Button onClick={handleActiveToggle} label="Open a dialog" />
-      <State store={store}>
-        <Dialog active={false} onEscKeyDown={handleActiveToggle} onOverlayClick={handleActiveToggle}>
-          <Banner color="mint" fullWidth icon={<IconCheckmarkMediumOutline />} onClose={handleActiveToggle}>
-            <Heading3>Succes: Dialog title</Heading3>
-          </Banner>
-          <Box padding={4}>
-            <TextBody>Here you can add arbitrary content.</TextBody>
-          </Box>
-          <ButtonGroup justifyContent="flex-end" padding={4}>
-            <Button label="Cancel" />
-            <Button label="Confirm" level="primary" />
-          </ButtonGroup>
-        </Dialog>
-      </State>
-    </Box>
-  ))
-  .add('warning style', () => (
-    <Box>
-      <Button onClick={handleActiveToggle} label="Open a dialog" />
-      <State store={store}>
-        <Dialog active={false} onEscKeyDown={handleActiveToggle} onOverlayClick={handleActiveToggle}>
-          <Banner color="gold" fullWidth icon={<IconWarningBadgedMediumOutline />} onClose={handleActiveToggle}>
-            <Heading3>Warning: Dialog title</Heading3>
-          </Banner>
-          <Box padding={4}>
-            <TextBody>Here you can add arbitrary content.</TextBody>
-          </Box>
-          <ButtonGroup justifyContent="flex-end" padding={4}>
-            <Button label="Cancel" />
-            <Button label="Confirm" level="primary" />
-          </ButtonGroup>
-        </Dialog>
-      </State>
-    </Box>
-  ))
-  .add('destructive style', () => (
-    <Box>
-      <Button onClick={handleActiveToggle} label="Open a dialog" />
-      <State store={store}>
-        <Dialog active={false} onEscKeyDown={handleActiveToggle} onOverlayClick={handleActiveToggle}>
-          <Banner color="ruby" fullWidth icon={<IconWarningBadgedMediumOutline />} onClose={handleActiveToggle}>
-            <Heading3>Succes: Dialog title</Heading3>
-          </Banner>
-          <ButtonGroup justifyContent="flex-end" padding={4}>
-            <Button label="Cancel" />
-            <Button label="Confirm" level="destructive" />
-          </ButtonGroup>
-        </Dialog>
-      </State>
-    </Box>
-  ));
+          </Dialog>
+        </State>
+      </Box>
+    );
+  });